test: remove node_modules mutation from modal harness
