+++ package.json
@@ -7,7 +7,7 @@
 	"scripts": {
 		"dev": "node esbuild.config.mjs",
 		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
-		"test": "node --experimental-test-module-mocks --import jiti/register --test src/services/match-url-parser.test.ts src/commands/create-match-note-from-url.test.ts src/commands/create-match-note-manually.test.ts src/commands/create-team-note.test.ts src/commands/create-player-note.test.ts src/commands/register-named-note-command.test.ts src/ui/match-url-modal-state.test.ts src/ui/named-note-modal.test.ts src/settings.test.ts src/team-player-note-schemas.test.ts src/services/match-note-files.test.ts src/services/team-player-note-files.test.ts src/services/match-note-template.test.ts src/services/team-player-note-template.test.ts src/services/note-title.test.ts src/services/match-note-workflow.test.ts src/services/team-player-note-workflow.test.ts src/services/manual-match-note-input.test.ts src/services/manual-match-note.test.ts src/services/manual-match-team-notes.test.ts src/services/user-facing-error.test.ts",
+		"test": "node --experimental-test-module-mocks --import ./src/ui/named-note-modal-test-preloader.mjs --import jiti/register --test src/services/match-url-parser.test.ts src/commands/create-match-note-from-url.test.ts src/commands/create-match-note-manually.test.ts src/commands/create-team-note.test.ts src/commands/create-player-note.test.ts src/commands/register-named-note-command.test.ts src/ui/match-url-modal-state.test.ts src/ui/named-note-modal.test.ts src/settings.test.ts src/team-player-note-schemas.test.ts src/services/match-note-files.test.ts src/services/team-player-note-files.test.ts src/services/match-note-template.test.ts src/services/team-player-note-template.test.ts src/services/note-title.test.ts src/services/match-note-workflow.test.ts src/services/team-player-note-workflow.test.ts src/services/manual-match-note-input.test.ts src/services/manual-match-note.test.ts src/services/manual-match-team-notes.test.ts src/services/user-facing-error.test.ts",
 		"prepare": "husky",
 		"format": "prettier . --write",
 		"format:check": "prettier . --check",
