+++ src/ui/named-note-modal.test.ts
@@ -1,131 +1,9 @@
-/* eslint-disable no-undef, @typescript-eslint/no-floating-promises */
+/* eslint-disable @typescript-eslint/no-floating-promises */
 
 import { deepStrictEqual, strictEqual } from 'node:assert';
-import { readFileSync, unlinkSync, writeFileSync } from 'node:fs';
-import { join } from 'node:path';
 import { setImmediate } from 'node:timers/promises';
-import { mock, test } from 'node:test';
-
-class TestElement {
-	children: TestElement[] = [];
-	listeners = new Map<string, Set<(event: KeyboardEvent) => void>>();
-	id = '';
-
-	empty(): void {
-		this.children = [];
-	}
-
-	createEl(): TestElement {
-		const child = new TestElement();
-		this.children.push(child);
-		return child;
-	}
-
-	createDiv(): TestElement {
-		return this.createEl();
-	}
-
-	addClass(): void {}
-
-	setAttr(): void {}
-
-	addEventListener(type: string, listener: (event: KeyboardEvent) => void): void {
-		const listeners = this.listeners.get(type) ?? new Set();
-		listeners.add(listener);
-		this.listeners.set(type, listeners);
-	}
-
-	removeEventListener(type: string, listener: (event: KeyboardEvent) => void): void {
-		this.listeners.get(type)?.delete(listener);
-	}
-
-	focus(): void {}
-
-	select(): void {}
-}
-
-class Modal {
-	contentEl = new TestElement();
-	closed = false;
-	title = '';
-
-	constructor(public app: unknown) {}
-
-	setTitle(title: string): void {
-		this.title = title;
-	}
-
-	close(): void {
-		this.closed = true;
-		(this as unknown as { onClose?: () => void }).onClose?.();
-	}
-}
-
-class ButtonComponent {
-	disabled = false;
-	private clickHandler: (() => void) | undefined;
-
-	constructor(public containerEl: TestElement) {}
-
-	setButtonText(): this {
-		return this;
-	}
-
-	setCta(): this {
-		return this;
-	}
-
-	onClick(handler: () => void): this {
-		this.clickHandler = handler;
-		return this;
-	}
-
-	setDisabled(disabled: boolean): this {
-		this.disabled = disabled;
-		return this;
-	}
-
-	click(): void {
-		if (!this.disabled) this.clickHandler?.();
-	}
-}
-
-class TextComponent {
-	inputEl = new TestElement();
-	private value = '';
-
-	constructor(public containerEl: TestElement) {}
-
-	setPlaceholder(): this {
-		return this;
-	}
-
-	setValue(value: string): this {
-		this.value = value;
-		return this;
-	}
-
-	getValue(): string {
-		return this.value;
-	}
-}
-
-const localObsidianPackagePath = join(process.cwd(), 'node_modules', 'obsidian', 'package.json');
-const localObsidianIndexPath = join(process.cwd(), 'node_modules', 'obsidian', 'index.js');
-const originalObsidianPackage = readFileSync(localObsidianPackagePath, 'utf8');
-const packageJson = JSON.parse(originalObsidianPackage) as { main?: string };
-packageJson.main = 'index.js';
-writeFileSync(localObsidianPackagePath, `${JSON.stringify(packageJson)}\n`);
-writeFileSync(localObsidianIndexPath, 'module.exports = {};\n');
-process.once('exit', () => {
-	writeFileSync(localObsidianPackagePath, originalObsidianPackage);
-	unlinkSync(localObsidianIndexPath);
-});
-
-mock.module('obsidian', {
-	namedExports: { ButtonComponent, Modal, TextComponent },
-	defaultExport: { ButtonComponent, Modal, TextComponent },
-});
+import { test } from 'node:test';
+import { ButtonComponent, TextComponent } from './named-note-modal-test-support';
 
 let NamedNoteModal: typeof import('./named-note-modal').NamedNoteModal;
 const namedNoteModalReady = import('./named-note-modal').then((module) => {
